Feat(client/style)/Fix(Client/Paginated): styled Products component, fixed broken Paginated component
