+++ src/components/Products/products.css
@@ -1,0 +1,32 @@
+.products__list{
+    margin-top: 1rem;
+    min-height: 25rem;
+}
+
+.products__list-item{
+    --item-rounded:15px;
+    display: flex;
+    flex-direction: row-reverse;
+    flex-wrap: wrap-reverse;
+    margin: .25rem;
+    padding: .25rem 1rem;
+    border-bottom-right-radius: var(--item-rounded);
+    border-top-left-radius: var(--item-rounded);
+    width: 100%;
+    justify-content: space-between;
+    background-color: rgba(var(--color-primary-rgb),.15);
+    transition: var(--transition);
+}
+
+.products__list-item:hover{
+    background-color: rgba(var(--color-primary-rgb),.30);
+}
+
+.products__list-item > p {
+    --p-rounded:5px;
+    padding: 0 .5rem;
+    border-bottom-right-radius: var(--p-rounded);
+    border-top-left-radius: var(--p-rounded);
+    margin: .25rem;
+    background-color: rgba(var(--color-primary-rgb),.15);
+}
